Refactored events.rs to be slightly less jumbo. It has wasted away to practically nothing.
