+++ src/handlers/mod.rs
@@ -3,6 +3,3 @@ pub mod api;
 pub mod history;
 pub mod file_ops;
 pub mod events;
-// TODO: Add these later when needed
-// pub mod input;
-// pub mod navigation;
